requirements.txt
selene == 2.0.0rc9
pytest == 8.3.5
allure-pytest == 2.14.3
python-dotenv == 1.1.1
selenium == 4.34.2
allure-python-commons == 2.14.3
